refactor: JwtProvider
