+++ src/main/java/com/synergy/backend/global/config/JwtConfig.java
@@ -1,0 +1,11 @@
+package com.synergy.backend.global.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+import com.synergy.backend.global.jwt.JwtProperties;
+
+@Configuration
+@EnableConfigurationProperties(JwtProperties.class)
+public class JwtConfig {
+}
